+++ packages/enhanced-img/src/index.js
@@ -1,12 +1,13 @@
 import path from 'node:path';
+import process from 'node:process';
 import { imagetools } from 'vite-imagetools';
 import { image } from './preprocessor.js';
 
 /**
- * @returns {Promise<import('vite').Plugin[]>}
+ * @returns {import('vite').Plugin[]}
  */
-export async function enhancedImages() {
-	const imagetools_instance = await imagetools_plugin();
+export function enhancedImages() {
+	const imagetools_instance = imagetools_plugin();
 	return !process.versions.webcontainer
 		? [image_plugin(imagetools_instance), imagetools_instance]
 		: [];
@@ -60,7 +61,7 @@ const fallback = {
 	'.webp': 'png'
 };
 
-async function imagetools_plugin() {
+function imagetools_plugin() {
 	/** @type {Partial<import('vite-imagetools').VitePluginOptions>} */
 	const imagetools_opts = {
 		defaultDirectives: async ({ pathname, searchParams: qs }, metadata) => {