+++ packages/kit/test/apps/basics/src/routes/read-file/+page.svelte
@@ -4,4 +4,6 @@
 
 <p data-testid="auto">{data.auto}</p>
 <p data-testid="url">{data.url}</p>
-<p data-testid="glob">{data.glob}</p>
+<p data-testid="local_glob">{data.local_glob}</p>
+<p data-testid="external_glob">{data.external_glob}</p>
+<div data-testid="svg">{@html data.svg}</div>