+++ packages/kit/test/apps/dev-only/test/test.js
@@ -2,6 +2,7 @@ import { expect } from '@playwright/test';
 import { test } from '../../../utils.js';
 import fs from 'node:fs';
 import path from 'node:path';
+import process from 'node:process';
 import { fileURLToPath } from 'node:url';
 
 const __dirname = path.dirname(fileURLToPath(import.meta.url));
@@ -107,14 +108,14 @@ test.describe('Vite', () => {
 		expect(manifest).toHaveProperty('optimized.e2e-test-dep-page-universal');
 	});
 
-	test('optimizes +page.server.js dependencies', async ({ page }) => {
+	test('skips optimizing +page.server.js dependencies', async ({ page }) => {
 		await page.goto('/');
 		await page.getByText('hello world!').waitFor();
 
 		const manifest_path = path.join(__dirname, '../node_modules/.vite/deps/_metadata.json');
 		const manifest = JSON.parse(fs.readFileSync(manifest_path, 'utf-8'));
 
-		expect(manifest).toHaveProperty('optimized.e2e-test-dep-page-server');
+		expect(manifest).not.toHaveProperty('optimized.e2e-test-dep-page-server');
 	});
 
 	test('optimizes +layout.svelte dependencies', async ({ page }) => {
@@ -137,14 +138,14 @@ test.describe('Vite', () => {
 		expect(manifest).toHaveProperty('optimized.e2e-test-dep-layout-universal');
 	});
 
-	test('optimizes +layout.server.js dependencies', async ({ page }) => {
+	test('skips optimizing +layout.server.js dependencies', async ({ page }) => {
 		await page.goto('/');
 		await page.getByText('hello world!').waitFor();
 
 		const manifest_path = path.join(__dirname, '../node_modules/.vite/deps/_metadata.json');
 		const manifest = JSON.parse(fs.readFileSync(manifest_path, 'utf-8'));
 
-		expect(manifest).toHaveProperty('optimized.e2e-test-dep-layout-server');
+		expect(manifest).not.toHaveProperty('optimized.e2e-test-dep-layout-server');
 	});
 
 	test('optimizes +error.svelte dependencies', async ({ page }) => {