+++ packages/migrate/migrations/self-closing-tags/index.js
@@ -1,5 +1,6 @@
 import colors from 'kleur';
 import fs from 'node:fs';
+import process from 'node:process';
 import prompts from 'prompts';
 import glob from 'tiny-glob/sync.js';
 import { remove_self_closing_tags } from './migrate.js';
@@ -10,7 +11,7 @@ export async function migrate() {
 	let compiler;
 	try {
 		compiler = await import_from_cwd('svelte/compiler');
-	} catch (e) {
+	} catch {
 		console.log(colors.bold().red('❌ Could not find a local Svelte installation.'));
 		return;
 	}
@@ -38,7 +39,7 @@ export async function migrate() {
 		try {
 			const code = await remove_self_closing_tags(compiler, fs.readFileSync(file, 'utf-8'));
 			fs.writeFileSync(file, code);
-		} catch (e) {
+		} catch {
 			// continue
 		}
 	}
@@ -48,9 +49,9 @@ export async function migrate() {
 }
 
 /** @param {string} name */
-async function import_from_cwd(name) {
+function import_from_cwd(name) {
 	const cwd = pathToFileURL(process.cwd()).href;
-	const url = await resolve(name, cwd + '/x.js');
+	const url = resolve(name, cwd + '/x.js');
 
 	return import(url);
 }